pretasks: split build script
pretasks is the first script getting executed by the build
process, then the actual build will happen and after that the
finalizing script will run.
this was needed to ensure the env gets properly prepared, all the
magic of /e/ will be there before actually starting building.
Last but not least the rest of the /e/ magic must be done like
delta builds etc.

Signed-off-by: steadfasterX <steadfasterX@gmail.com>