postsync: removed OTA URL replacement
I tried to make that work with the suggested way of an
overlay but was not able to overwrite the ecloud one.

Removed that code as replacing the source for the OTA URI is not
recommended anyways (see LineageOTA README).

A reliable and always working way is setting the OTA URI in a system property
override in your device tree instead. e.g:

> # OTA server
> PRODUCT_PROPERTY_OVERRIDES += \
>    lineage.updater.uri=https://<your-ota-server/api/v1/{device}/{type}/{incr}

(note: on Nougat and lower it must be cm.updater.uri instead)

{device}, {type} and {incr} are macros so do not replace these there.
The correct URI depends on your OTA server setp ofc.

Signed-off-by: steadfasterX <steadfasterX@gmail.com>