complete behavior re-write
mka eos-fresh is leagcy now. It was never a good idea to include
that but now after I had the time to test everything I found it
will never work as expected. so bye bye eos-fresh.

That makes `mka eos` the only build target as of now.

Other then that I also found issues with the mka eos approach and
here as well I had just taken over parts of the docker scripts to
keep it simple. Most importantly: breakfast was included which has
been removed completely now.

Almost all magic scripts are executed (depending on the vars)
directly by vendor/e/vendorsetup.sh now as this is the only way to
ensure that we get all prepared as needed (which is too late in
when using a build target approach).

From now on you HAVE TO either LUNCH or BREAKFAST your device
so exactly as you should have expected it from the beginning.

Example:

```
source build/envsetup.sh
breakfast h815
mka eos
```

VARIABLES:

 - new: `EOS_CLEAN_BEFORE_BUILD` (mka clean before building)
 - renamed: `CLEAN_OUTDIR` was misleading and moved to CLEAN_ZIPDIR
   and `EOS_CLEAN_ZIPDIR` instead

Not that important: all special /e/ steps are now better visible
during the build process.

Signed-off-by: steadfasterX <steadfasterX@gmail.com>