vendorsetup: fix reset and dup vars
the reset is needed to ensure no environment variables from a
previous run will be used for a new run.
e.g when you wanted to clean before build on a previous but not
for the next and the variable is still there it would be cleaned
always.
the reset was unfortunately not working correctly before which
has been fixed now.

also:

- removing duplicated var
- creating keys dir when needed only

Signed-off-by: steadfasterX <steadfasterX@gmail.com>